+++ samples/CarterAndMVC/CarterAndMVC.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <AssemblyName>CarterAndMVC</AssemblyName>
     <OutputType>Exe</OutputType>
     <LangVersion>latest</LangVersion>
