+++ samples/CarterSample/Features/Directors/DirectorsModule.cs
@@ -1,7 +1,5 @@
-namespace CarterSample.Directors;
+namespace CarterSample.Features.Directors;
 
-using System.Diagnostics.CodeAnalysis;
-using Carter.OpenApi;
 using FluentValidation;
 using Microsoft.AspNetCore.Http.HttpResults;
 
@@ -56,11 +54,6 @@ public class DirectorsModule : CarterModule
                 return TypedResults.NoContent();
             });
     }
-
-    private static async Task<Results<NoContent, NotFound, Ok<string>>> Handler(HttpContext context)
-    {
-        return TypedResults.Ok("");
-    }
 }
 
 public class Person
@@ -81,52 +74,7 @@ public class PersonValidation : AbstractValidator<Person>
     }
 }
 
-interface IUpdatePersonCommand
+internal interface IUpdatePersonCommand
 {
     bool Execute(Person person);
-}
-
-public static class RouteExtensions
-{
-    public static RouteHandlerBuilder MapPost<T>(
-        this IEndpointRouteBuilder endpoints,
-        [StringSyntax("Route")] string pattern,
-        Delegate handler)
-    {
-        return endpoints.MapPost(pattern, handler).AddRouteHandlerFilter(async (context, next) =>
-        {
-            var tdparam = context.GetArgument<T>(0);
-
-            var result = context.HttpContext.Request.Validate(tdparam);
-            if (!result.IsValid)
-            {
-                var problemDetailsErrors = new Dictionary<string, object>()
-                    { { "errors", result.GetFormattedErrors() } };
-                return Results.Problem(statusCode: 422, extensions: problemDetailsErrors);
-            }
-
-            return await next(context);
-        });
-    }
-
-    public static RouteHandlerBuilder MapPut<T>(
-        this IEndpointRouteBuilder endpoints,
-        [StringSyntax("Route")] string pattern,
-        Delegate handler)
-    {
-        return endpoints.MapPut(pattern, handler).AddRouteHandlerFilter(async (context, next) =>
-        {
-            var tdparam = context.GetArgument<T>(0);
-
-            var result = context.HttpContext.Request.Validate(tdparam);
-            if (!result.IsValid)
-            {
-                var problemDetailsErrors = new Dictionary<string, object>()
-                    { { "errors", result.GetFormattedErrors() } };
-                return Results.Problem(statusCode: 422, extensions: problemDetailsErrors);
-            }
-
-            return await next(context);
-        });
-    }
 }
