+++ template/content/HomeModule.cs
@@ -1,13 +1,4 @@
-namespace CarterTemplate
+﻿public class HomeModule : ICarterModule
 {
-    using Carter;
-    using Microsoft.AspNetCore.Http;
-
-    public class HomeModule : CarterModule
-    {
-        public HomeModule()
-        {
-            Get("/", async(req, res) => await res.WriteAsync("Hello from Carter!"));
-        }
-    }
+    public void AddRoutes(IEndpointRouteBuilder app) => app.MapGet("/", () => "Hello from Carter!");
 }
