+++ test/Carter.ResponseNegotiators.Newtonsoft.Tests/Carter.ResponseNegotiators.Newtonsoft.Tests.csproj
@@ -1,8 +1,5 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-    </PropertyGroup>
 
 
     <ItemGroup>
