+++ test/Carter.Tests/Carter.Tests.csproj
@@ -1,7 +1,4 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="fakeiteasy" Version="6.0.0" />
   </ItemGroup>
